documentation of smart contract

+++ contracts/CovidCredentialRegistry.sol
@@ -7,7 +7,7 @@ contract CovidCredentialRegistry is ICredentialRegistry, WhitelistedRole {
 
   mapping (bytes32 => mapping (address => CovidMetadata)) credentials;
 
-  function register(bytes32 hash, bytes32 id, uint exp, bool sex, uint8 age, bytes32 ubigeo, uint16 zipcode, CovidCode credentialType, InterruptionReason reason) override external returns(bool) {
+  function register(bytes32 hash, bytes32 id, uint exp, bool sex, uint8 age, bytes32 ubigeo, uint32 zipcode, CovidCode credentialType, InterruptionReason reason) override external returns(bool) {
     require(super.isWhitelisted(msg.sender), "Account isn't whitelisted");
     CovidMetadata storage credential = credentials[hash][msg.sender];
     require(credential.id==0,"Credential ID already exists");
